+++ AirportBrcmFixup/kern_fakebrcm.cpp
@@ -10,6 +10,33 @@ OSDefineMetaClassAndStructors(FakeBrcm, IOService);
 IOService *FakeBrcm::service_provider {nullptr};
 OSDictionary *FakeBrcm::service_dict {nullptr};
 OSDictionary *FakeBrcm::prop_table {nullptr};
+FakeBrcm::t_config_read16  FakeBrcm::orgConfigRead16 {nullptr};
+FakeBrcm::t_config_read32  FakeBrcm::orgConfigRead32 {nullptr};
+
+//==============================================================================
+
+int getIntegerProperty(IORegistryEntry* entry, const char *aKey)
+{
+    OSData* data = OSDynamicCast(OSData, entry->getProperty(aKey));
+    if (!data || sizeof(UInt32) != data->getLength())
+    {
+        return -1;
+    }
+    UInt32 result = *static_cast<const UInt32*>(data->getBytesNoCopy());
+    return result;
+}
+
+//==============================================================================
+
+bool isServiceSupported(IOService* service)
+{
+    for (int i=0; i<kextListSize; i++)
+    {
+        if (strcmp(serviceNameList[i], service->getName()) == 0)
+            return true;
+    }
+    return false;
+}
 
 //==============================================================================
 
@@ -40,6 +67,16 @@ bool FakeBrcm::init(OSDictionary *propTable)
 
 //==============================================================================
 
+bool FakeBrcm::attach(IOService *provider)
+{  
+    if (!hookProvider(provider))
+        return false;
+    
+    return super::attach(provider);
+}
+
+//==============================================================================
+
 IOService* FakeBrcm::probe(IOService * provider, SInt32 *score)
 {
     DBGLOG("BRCMFX @ FakeBrcm::probe()");
@@ -59,6 +96,7 @@ IOService* FakeBrcm::probe(IOService * provider, SInt32 *score)
     }
     
     service_provider = provider;
+    DBGLOG("BRCMFX @ FakeBrcm::probe(): service provider is %s", provider->getName());
     
     for (int i=0; i<kextListSize; i++)
     {
@@ -81,7 +119,7 @@ IOService* FakeBrcm::probe(IOService * provider, SInt32 *score)
     
     if (service_dict->getCount() != 0)
     {
-        DBGLOG("BRCMFX @ FakeBrcm::probe() will change score from %d to 1300", *score);
+        DBGLOG("BRCMFX @ FakeBrcm::probe() will change score from %d to 2000", *score);
         *score = 2000;  // change probe score to be the first in the list
     }
     else
@@ -117,6 +155,9 @@ bool FakeBrcm::start(IOService *provider)
         SYSLOG("BRCMFX @ FakeBrcm::start(): fallback to original driver");
         return false;
     }
+    
+    if (!hookProvider(provider))
+        return false;
 
     return true;
 }
@@ -126,6 +167,8 @@ bool FakeBrcm::start(IOService *provider)
 void FakeBrcm::stop(IOService *provider)
 {
     DBGLOG("BRCMFX @ FakeBrcm::stop()");
+    
+    unhookProvider();
 
     super::stop(provider);
 }
@@ -137,6 +180,8 @@ void FakeBrcm::free()
     DBGLOG("BRCMFX @ FakeBrcm::free()");
     
     OSSafeReleaseNULL(service_dict);
+    
+    unhookProvider();
 
     super::free();
 }
@@ -168,4 +213,108 @@ IOService* FakeBrcm::getService(const char* service_name)
     return service;
 }
 
+//==============================================================================
 
+UInt16 FakeBrcm::configRead16(IOService *that, UInt32 space, UInt8 offset)
+{
+    UInt16 result = orgConfigRead16(that, space, offset);
+    UInt16 newResult = result;
+    
+    if (that == service_provider || isServiceSupported(that))
+    switch (offset)
+    {
+        case kIOPCIConfigVendorID:
+        {
+            int vendor = getIntegerProperty(that, "vendor-id");
+            if (-1 != vendor)
+                newResult = vendor;
+            break;
+        }
+        case kIOPCIConfigDeviceID:
+        {
+            int device = getIntegerProperty(that, "device-id");
+            if (-1 != device)
+                newResult = device;
+            break;
+        }
+    }
+    
+    if (newResult != result)
+        DBGLOG("BRCMFX @ FakeBrcm::configRead16: name = %s, source value = 0x%04x replaced with value = 0x%04x", that->getName(), result, newResult);
+
+    return newResult;
+}
+
+//==============================================================================
+
+UInt32 FakeBrcm::configRead32(IOService *that, UInt32 space, UInt8 offset)
+{
+    UInt32 result = orgConfigRead32(that, space, offset);
+    UInt32 newResult = result;
+    
+    if (that == service_provider || isServiceSupported(that))
+    switch (offset)
+    {
+        case kIOPCIConfigVendorID:
+        case kIOPCIConfigDeviceID: // OS X does a non-aligned read, which still returns full vendor / device ID
+        {
+            int vendor = getIntegerProperty(that, "vendor-id");
+            if (-1 != vendor)
+                newResult = (newResult & 0xFFFF0000) | vendor;
+            
+            int device = getIntegerProperty(that, "device-id");
+            if (-1 != device)
+                newResult = (device << 16) | (newResult & 0xFFFF);
+            break;
+        }
+    }
+    
+    if (newResult != result)
+        DBGLOG("BRCMFX @ FakeBrcm::configRead32: name = %s, source value = 0x%08x replaced with value = 0x%08x", that->getName(), result, newResult);
+    
+    return newResult;
+}
+
+//==============================================================================
+
+bool FakeBrcm::hookProvider(IOService *provider)
+{
+    void * pcidev = static_cast<void *>(provider);
+    uint64_t * vmt = pcidev ? static_cast<uint64_t **>(pcidev)[0] : nullptr;
+    if (vmt && vmt[VMTOffset::configRead16] != reinterpret_cast<uint64_t>(FakeBrcm::configRead16))
+    {
+        orgConfigRead16 = reinterpret_cast<t_config_read16>(vmt[VMTOffset::configRead16]);
+        vmt[VMTOffset::configRead16] = reinterpret_cast<uint64_t>(FakeBrcm::configRead16);
+        DBGLOG("BRCMFX @ FakeBrcm::hookProvider for configRead16 was successful");
+    }
+    
+    if (vmt && vmt[VMTOffset::configRead32] != reinterpret_cast<uint64_t>(FakeBrcm::configRead32))
+    {
+        orgConfigRead32 = reinterpret_cast<t_config_read32>(vmt[VMTOffset::configRead32]);
+        vmt[VMTOffset::configRead32] = reinterpret_cast<uint64_t>(FakeBrcm::configRead32);
+        DBGLOG("BRCMFX @ FakeBrcm::hookProvider for configRead32 was successful");
+    }
+    return true;
+}
+
+//==============================================================================
+
+void FakeBrcm::unhookProvider()
+{
+    if (service_provider != nullptr && orgConfigRead16 != nullptr)
+    {
+        void * pcidev = static_cast<void *>(service_provider);
+        uint64_t * vmt = pcidev ? static_cast<uint64_t **>(pcidev)[0] : nullptr;
+        if (vmt && vmt[VMTOffset::configRead16] != reinterpret_cast<uint64_t>(orgConfigRead16))
+        {
+            vmt[VMTOffset::configRead16] = reinterpret_cast<uint64_t>(orgConfigRead16);
+            DBGLOG("BRCMFX @ FakeBrcm::unhookProvider for configRead16 was successful");
+        }
+        
+        if (vmt && vmt[VMTOffset::configRead32] != reinterpret_cast<uint64_t>(orgConfigRead32))
+        {
+            vmt[VMTOffset::configRead32] = reinterpret_cast<uint64_t>(orgConfigRead32);
+            DBGLOG("BRCMFX @ FakeBrcm::unhookProvider for configRead32 was successful");
+        }
+    }
+}