+++ AirportBrcmFixup/kern_start.cpp
@@ -25,9 +25,10 @@ const char *Configuration::bootargBeta[] {
 	"-brcmfxbeta"
 };
 
-const char *Configuration::bootargBrcmCountry   {"brcmfx-country"};     // 5 Ghz patch - change default country
-const char *Configuration::bootargWlanMsgLevel  {"wl_msg_level"};       // message level for wlan kexts
-const char *Configuration::bootargWlanMsgLevel2 {"wl_msg_level2"};      // message level for wlan kexts
+const char *Configuration::bootargBrcmCountry    {"brcmfx-country"};     // 5 Ghz patch - change default country
+const char *Configuration::bootargWlanMsgLevel   {"wl_msg_level"};       // message level for wlan kexts
+const char *Configuration::bootargWlanMsgLevel2  {"wl_msg_level2"};      // message level for wlan kexts
+const char *Configuration::bootargBrcmEnableWowl {"-brcmfxwowl"};        // enable WOWL 
 
 
 Configuration config;
@@ -36,6 +37,7 @@ Configuration config;
 void Configuration::readArguments() {
     
     country_code_overrided = PE_parse_boot_argn(bootargBrcmCountry, country_code, sizeof(country_code));
+
     
     if (!PE_parse_boot_argn(bootargWlanMsgLevel, &wl_msg_level, sizeof(wl_msg_level)))
         wl_msg_level = 0;
@@ -44,6 +46,7 @@ void Configuration::readArguments() {
     
     char tmp[16];
     disabled = PE_parse_boot_argn(config.bootargOff[0], tmp, sizeof(tmp));
+    enable_wowl = PE_parse_boot_argn(bootargBrcmEnableWowl, tmp, sizeof(tmp));
 }